Create DWSIM_Exploit.py
This is an exploit script for the DWSIM simulation. It reads from the first 4 registers and writes to the 10th register. The first 3 indicate the levels of water, nitrogen and acetone respectively. The fourth register indicates the current temperature to which the water is heated to before being passed into the absorption column.

The 10th register at address 0x9 is the way in which the HMI is able to change the temperature of the heater. This will be the pathway through which we will exploit.